+++ src/lib/utils/api-response.ts
@@ -1,5 +1,5 @@
 import { CLIENT_ERROR_MESSAGES } from '@/constants/error-messages';
-import type { ApiResponse } from '@/types/api';
+import type { ApiResponse } from '@/types';
 
 interface FetchOptions extends RequestInit {
   url: string;
